+++ internal/xerrors/stacktrace.go
@@ -6,6 +6,8 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/stack"
 )
 
+const KeywordAt = "at"
+
 type withStackTraceOptions struct {
 	skipDepth int
 }
@@ -51,7 +53,7 @@ type stackError struct {
 }
 
 func (e *stackError) Error() string {
-	return e.err.Error() + " at `" + e.stackRecord + "`"
+	return e.err.Error() + " " + KeywordAt + " `" + e.stackRecord + "`"
 }
 
 func (e *stackError) Unwrap() error {