+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed session closing in `ydb.WithExecuteDataQueryOverQueryClient(true)` scenario
+
 ## v3.110.1
 * Added the ability to send BulkRequest exceeding the GrpcMaxMessageSize