+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added the previously lost HintSessionBalancer
+
 ## v3.110.1
 * Added the ability to send BulkRequest exceeding the GrpcMaxMessageSize