+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* Pinged new connections on discovery attempt, closed dropped ones, so `ydb_go_sdk_ydb_driver_conns` metric is correct
+* Fixed `ydb_go_sdk_ydb_driver_conns` metric to correctly display nodes that are discovered at the moment
 
 ## v3.110.1
 * Added the ability to send BulkRequest exceeding the GrpcMaxMessageSize