+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added `PrintErrorWithoutStack` helper for remove stack records from error string
+
 ## v3.110.1
 * Added the ability to send BulkRequest exceeding the GrpcMaxMessageSize