+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Added `sugar.PrintErrorWithoutStack` helper for remove stack records from error string
+* Added `sugar.UnwrapError` helper for unwrap source error to root errors
 
 ## v3.110.1
 * Added the ability to send BulkRequest exceeding the GrpcMaxMessageSize