+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Replaced requirements from go1.23 to go1.22
+
 ## v3.111.0
 * Added `sugar.PrintErrorWithoutStack` helper for remove stack records from error string
 * Added `sugar.UnwrapError` helper for unwrap source error to root errors